streaming-prototype/scorer/requirements.txt
kafka-python==2.0.2
duckdb>=0.10.0
